version for master 0.3
