+++ src/main/java/neointernship/web/server/Server.java
@@ -2,27 +2,7 @@ package neointernship.web.server;
 
 import neointernship.chess.game.model.enums.ChessType;
 import neointernship.chess.game.model.enums.Color;
-import neointernship.chess.game.model.enums.EnumGameState;
-import neointernship.chess.game.model.enums.EnumMatchResult;
-import neointernship.chess.game.model.figure.factory.Factory;
-import neointernship.chess.game.model.figure.factory.IFactory;
-import neointernship.chess.game.model.figure.piece.Figure;
-import neointernship.chess.game.model.mediator.IMediator;
-import neointernship.chess.game.model.mediator.Mediator;
-import neointernship.chess.game.model.playmap.board.Board;
-import neointernship.chess.game.model.playmap.board.IBoard;
-import neointernship.chess.game.model.playmap.board.figuresstartposition.FiguresStartPositionRepository;
-import neointernship.chess.game.model.playmap.field.Field;
-import neointernship.chess.game.model.playmap.field.IField;
-import neointernship.chess.game.story.IStoryGame;
-import neointernship.chess.game.story.StoryGame;
 import neointernship.chess.logger.ErrorLoggerServer;
-import neointernship.chess.logger.GameLogger;
-import neointernship.chess.statistics.Statistics;
-import neointernship.web.client.communication.data.endgame.EndGame;
-import neointernship.web.client.communication.data.endgame.IEndGame;
-import neointernship.web.client.communication.data.initgame.IInitGame;
-import neointernship.web.client.communication.data.initgame.InitGame;
 import neointernship.web.client.communication.data.initinfo.InitInfoDto;
 import neointernship.web.client.communication.message.ClientCodes;
 import neointernship.web.client.communication.message.IMessage;
@@ -38,7 +18,6 @@ import java.net.BindException;
 import java.net.ServerSocket;
 import java.net.Socket;
 import java.text.SimpleDateFormat;
-import java.time.LocalTime;
 import java.util.Queue;
 import java.util.concurrent.ConcurrentLinkedQueue;
 
@@ -71,238 +50,9 @@ public class Server implements IServer {
         lobbyList = new ConcurrentLinkedQueue<>();
     }
 
-    private class Lobby extends Thread {
-        private final IBoard board;
-        private final IFactory figureFactory;
-        private final IMediator mediator;
-        private final IPossibleActionList possibleActionList;
-        private final IStoryGame storyGame;
 
-        private final int lobbyId;
-        private final ChessType chessTypes;
-        private final FiguresStartPositionRepository figuresStartPositionRepository;
-        private final Character FIELD_CHAR_EMPTY = '.';
 
-        private final IGameLoop gameLoop;
 
-        private final Server server;
-        private final ActiveConnectionController connectionController;
-        private ActiveColorController activeColorController;
-
-        private Lobby(final UserConnection firstUserConnection, final UserConnection secondUserConnection,
-                      final int lobbyId, final Server server, final ChessType chessType) {
-            this.lobbyId = lobbyId;
-            this.server = server;
-
-            activeColorController = new ActiveColorController();
-            this.connectionController = new ActiveConnectionController(firstUserConnection, secondUserConnection,
-                    (IColorControllerSubscriber) activeColorController);
-
-            board = new Board();
-            figureFactory = new Factory();
-            mediator = new Mediator();
-            storyGame = new StoryGame(mediator);
-            possibleActionList = new PossibleActionList(board, mediator, storyGame);
-
-            this.chessTypes = chessType;
-            figuresStartPositionRepository = new FiguresStartPositionRepository();
-
-            //TODO:
-            GameLogger.addLogger(lobbyId);
-
-            gameLoop = new GameLoop(mediator, possibleActionList, board, activeColorController, storyGame);
-
-            GameLogger.getLogger(lobbyId).logStartGame(firstUserConnection.getName(), secondUserConnection.getName());
-
-            initGameMap();
-        }
-
-        private void initGameMap() {
-            final Character[][] figuresRepository = figuresStartPositionRepository.getStartPosition(chessTypes);
-            for (int i = 0; i < board.getSize(); i++) {
-                for (int j = 0; j < board.getSize(); j++) {
-                    final IField field = board.getField(i, j);
-
-                    final Character currentChar = figuresRepository[i][j];
-                    if (currentChar != FIELD_CHAR_EMPTY) {
-                        final Color color = i > 4 ? Color.WHITE : Color.BLACK;
-                        final Figure figure = figureFactory.createFigure(currentChar, color);
-
-                        mediator.addNewConnection(field, figure);
-                    }
-                }
-            }
-            possibleActionList.updateRealLists();
-        }
-
-        private void sendInitInfo() {
-            for (final UserConnection userConnection : connectionController.getConnections()) {
-                final BufferedWriter out = userConnection.getOut();
-                final IMessage msg = new Message(ClientCodes.INIT_GAME);
-                final IInitGame initGame = new InitGame(mediator, board, userConnection.getColor());
-
-                try {
-                    send(out, MessageSerializer.serialize(msg));
-                    send(out, InitGameSerializer.serialize(initGame));
-                } catch (final IOException e) {
-                    ErrorLoggerServer.logException(e);
-                }
-            }
-        }
-
-        private void sendUpdatedMediator(final IAnswer answer, final TurnStatus turnStatus) {
-            for (final UserConnection userConnection : connectionController.getConnections()) {
-                final BufferedWriter out = userConnection.getOut();
-                final IMessage msg = new Message(ClientCodes.UPDATE);
-                final IUpdate update = new Update(answer, turnStatus);
-
-                try {
-                    send(out, MessageSerializer.serialize(msg));
-                    send(out, UpdateSerializer.serialize(update));
-                } catch (final IOException e) {
-                    ErrorLoggerServer.logException(e);
-                }
-            }
-        }
-
-        private void sendEndGame(final EnumGameState enumGameState) {
-            for (final UserConnection userConnection : connectionController.getConnections()) {
-                final BufferedWriter out = userConnection.getOut();
-                final IMessage msg = new Message(ClientCodes.END_GAME);
-
-                final IEndGame endGame = new EndGame(enumGameState);
-
-                try {
-                    send(out, MessageSerializer.serialize(msg));
-                    send(out, EndGameSerializer.serialize(endGame));
-                } catch (final IOException e) {
-                    ErrorLoggerServer.logException(e);
-                }
-            }
-        }
-
-        private IAnswer transformation(final IAnswer answer, final TurnStatus turnStatus,
-                                       final BufferedReader in, final BufferedWriter out) throws IOException {
-            sendUpdatedMediator(answer, turnStatus);
-
-            final Message message1 = new Message(ClientCodes.TRANSFORMATION);
-            send(out, MessageSerializer.serialize(message1));
-
-            String string = in.readLine();
-            TransformationDto transformationDto = TransformationSerializer.deserialize(string);
-            char symbol = transformationDto.getFigureChar();
-
-            return new AnswerSimbol(answer.getFinalX(), answer.getFinalY(),
-                    answer.getFinalX(), answer.getFinalY(), symbol);
-
-        }
-
-        @Override
-        public void run() {
-            sendInitInfo();
-            final ConsoleBoardWriter boardWriter = new ConsoleBoardWriter(mediator, board);
-            MessageDto answerMsg = null;
-            UserConnection connection = null;
-
-            while (gameLoop.isAlive()) {
-                connectionController.update();
-                connection = connectionController.getCurrentConnection();
-
-                TurnStatus turnStatus = null;
-                IAnswer answer = null;
-
-                do {
-                    try {
-                        final BufferedWriter out = connection.getOut();
-                        final IMessage message = new Message(ClientCodes.TURN);
-                        send(out, MessageSerializer.serialize(message));
-
-                        final BufferedReader in = connection.getIn();
-                        final String msg = in.readLine();
-                        answerMsg = MessageSerializer.deserialize(msg);
-
-                        if (answerMsg.getClientCodes() == ClientCodes.END_GAME) break;
-
-                        final String turnString = in.readLine();
-                        final TurnDto turnDto = AnswerSerializer.deserialize(turnString);
-                        turnDto.validate();
-                        answer = turnDto.getAnswer();
-
-                        turnStatus = makeTurn(answer);
-
-                        if (turnStatus == TurnStatus.TRANSFORMATION_BEFORE) {
-                            answer = transformation(answer, turnStatus, in, out);
-                            turnStatus = makeTurn(answer);
-                        }
-
-                        GameLogger.getLogger(lobbyId).logPlayerMoveAction(connection.getColor(),
-                                mediator.getFigure(new Field(answer.getFinalX(), answer.getFinalY())),
-                                new Field(answer.getStartX(), answer.getStartY()),
-                                new Field(answer.getFinalX(), answer.getFinalY()), turnStatus);
-                    } catch (final Exception e) {
-                        ErrorLoggerServer.logException(e);
-                    }
-
-                } while (turnStatus == TurnStatus.ERROR);
-                if (answerMsg.getClientCodes() == ClientCodes.END_GAME) break;
-
-                sendUpdatedMediator(answer, turnStatus);
-                boardWriter.printBoard();
-            }
-
-            gameLoop.getMatchResult();
-
-            EnumGameState enumGameState = gameLoop.getMatchResult().getValue();
-
-            if (answerMsg.getClientCodes() == ClientCodes.END_GAME) enumGameState = EnumGameState.RESIGNATION;
-
-            sendEndGame(enumGameState);
-            GameLogger.getLogger(lobbyId).logEndGame(enumGameState);
-
-            setStatistics(connection, enumGameState);
-            downService();
-        }
-
-        public void setStatistics(UserConnection connection, final EnumGameState enumGameState) {
-            final EnumMatchResult matchResult = EnumMatchResult.getEnumMatchResult(enumGameState);
-            final String nameFirstPlayer = connection.getName();
-            final Color colorFirstPlayer = connection.getColor();
-
-            connectionController.update();
-            connection = connectionController.getCurrentConnection();
-
-            final String nameSecondPlayer = connection.getName();
-            final Color colorSecondPlayer = connection.getColor();
-            try {
-                Statistics.setStatistics(nameFirstPlayer, colorFirstPlayer, matchResult,
-                        nameSecondPlayer, colorSecondPlayer, EnumMatchResult.swapEnumMatchResult(matchResult));
-            } catch (final Exception exception) {
-                exception.printStackTrace();
-            }
-        }
-
-        public TurnStatus makeTurn(final IAnswer answer) {
-            return gameLoop.doIteration(answer);
-        }
-
-        /**
-         * закрытие сервера, удаление себя из списка нитей
-         */
-        private void downService() {
-            for (int i = 0; i < 2; i++) {
-                connectionController.update();
-                try {
-                    final UserConnection currentConnection = connectionController.getCurrentConnection();
-                    if (!currentConnection.getSocket().isClosed()) {
-                        currentConnection.getSocket().close();
-                        currentConnection.getOut().close();
-                        currentConnection.getIn().close();
-                        server.lobbyList.remove(this);
-                    }
-                } catch (final IOException ignored) {
-                }
-            }
-        }
     public void deleteLobby(final Lobby lobby) {
         lobbyList.remove(lobby);
     }
@@ -451,17 +201,7 @@ public class Server implements IServer {
         }
     }
 
-    /**
-     * отсылка одного сообщения клиенту
-     *
-     * @param msg сообщение
-     */
-    private void send(final BufferedWriter out, final String msg) throws IOException {
-        out.write(msg + "\n");
-        out.flush();
-    }
-
-    public static void main(final String[] args) throws IOException {
+    public static void main(final String[] args) {
         final Server server = new Server();
         server.startServer();
     }
